Upgrade to Dalamud.NET.Sdk 15.0.0, fix TerritoryChanged signature

+++ SpotifyHonorific/Updaters/Updater.cs
@@ -100,7 +100,7 @@ public class Updater : IDisposable
         GC.SuppressFinalize(this);
     }
 
-    private void OnTerritoryChanged(ushort territoryId)
+    private void OnTerritoryChanged(uint territoryId)
     {
         _titleState.ForceResend();
     }
